Persistence Bug Fix.
Previous: Showed a negative Fibonacci number when the data type (long)
was out of range. Reading of persistence data was incorrect if the
Server Service was started with the OSGi configuration property set to
(1), instead of the default value (4).

Now: Reading persistence data is corrected. Odd or even number of
Fibonacci sequence size, will not affect the value of "SecondLastFib" or
"LastFib" (long). The order of the Fibonacci number is ascending,
meaning "LastFib" is always greater than "SecondLastFib".